Prestige Ocean Pearl Kozhikode Kerala
Prestige Ocean Pearl Kozhikode is a residential accommodation that welcomes you in pure bliss. It is a place which is complemented by best facilities and world class infrastructure in a well equipped community. Illustration of the finest creation inspired by vibrant artistic cultures. 

Prestige Ocean Pearl offers spacious and comfortable private living space to improve your daily life. Offering 1 BHK, 2 BHK and 3 BHK apartments in Kozhikode offer complementary lifestyle convenience. 

These are exquisitely picturesque surroundings, spacious apartments, gracious residences with stellar amenities where you can connect with like-minded community. This makes it one of the most promising and luxurious real estate projects in Kerala.

Prestige Ocean Pearl Calicut is beautiful in design, excellent in location and high end in amenities. Offering you the bliss of a lifetime with the panoramic views of the beautiful city as you enjoy peace and quiet. Don't wait for the right time, buy a home here and make the right time now. Let's wait for Srishti's spark to hit.

Major schools and colleges, reputed hospitals and shopping destinations are close by. A home to live life that will never be short of drama and which is perfectly designed to inspire the artist in you. Filled with unique lifestyle features like a multi-purpose cinema screen with an amphitheater, a mural wall, a piano-themed passageway and a swimming pool shaped like a guitar headstock, this is one of the most premium residential projects in Kozhikode Kerala.

So, if you are looking for an escape from the everyday chaos of city life or you are an artist looking for your collection at these ideal place to stay in this residential project in Kozhikode Kerala.

For More Details
Visit us - https://bit.ly/3ytYwoH